Add one configuration for developers to easily ignore index (#317)
## What changes were proposed in this pull request?
This option is for OAP developers to easily analyze the OAP
performance w/ and w/o index even if the index is created already.

Usage:
sqlContext.conf.setConfString(SQLConf.OAP_USE_INDEX.key, value)

The value is true by default. This means that the index will be
used by default. Otherwise, the index won't be used.

## How was this patch tested?
mvn test

* Add one configuration for developers to easily ignore index

This option is for OAP developers to easily analyze the OAP
performance w/ and w/o index even if the index is created already.

Usage:
sqlContext.conf.setConfString(SQLConf.OAP_USE_INDEX.key, value)

The value is true by default. This means that the index will be
used by default. Otherwise, the index won't be used.

* No available index if OAP_USE_INDEX_FOR_DEVELOPERS is false

1. replace OAP_USE_INDEX with OAP_USE_INDEX_FOR_DEVELOPERS
2. hasAvailableIndex will report false to strategies if the option
   is false.

* Fix the style errors

* Fix the checking errors

* Be more concise

* Correct the indention

* Add one unit test

The test case is to test the iterator size of oap data file for one
specific filter.
If using index, the returned iterator size is already filtered by index.
Otherwise, it's the count of the total size.

* Follow the correct style.