[Orc-support-02] Add four orc record readers (#872)
* Add four orc record readers

The four orc readers are below:

OrcColumnarBatchReader.java and extended
IndexedOrcColumnarBatchReader.java

OrcMapreduceRecordReader.java and extended
IndexedOrcMapreduceRecordReader.java

OrcColumnarBatchReader is a copy of OrcColumnarBatchReader
from Spark2.3 with minor changes to be able to have a child class
IndexedOrcColumnarBatchReader.

OrcMapreduceRecordReader is a copy of OrcMapreduceRecordReader
from orc upstream with minor changes to be able to have a child class
IndexedOrcMapreduceRecordReader.

The above four readers are specifically used by oap to enable orc file
format reader to use oap index.

* Replace IntStream with for loop

The IntStream is the JDK1.8 feature.
Spark2.1 is using 1.7. Thus change back to for loop.

* Refactoring the code

Refactoring the code as follows to keep the concise code style.
1. Correct the indentation
2. Add more comments
3. Extract the common code
4. a little bit performance tuning for indexed readers.

* Change the dependent column utils package path with extra oap.orc

The dependent column utils are columnar batch, columnar vectors and
the dependent column classes which are only targeted for orc columnar
batch readers.

Previously, the package path keeps the same as Spark 2.3, although it's
already different from Spark 2.1 and 2.2.

In order to explicitly distinguish from the existing columnar batch and
dependent column utils in Spark 2.1 and 2.2, now the package patch is changed
from "org.apache.spark.sql.vectorized"
to "org.apache.spark.sql.vectorized.oap.orc".